management-plane/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
pydantic==2.9.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
